+++ jsonargparse_tests/base.py
@@ -2,11 +2,13 @@ import inspect
 import os
 import platform
 import shutil
+import sys
 import tempfile
 import unittest
 from contextlib import contextmanager, redirect_stderr
 from importlib.util import find_spec
 from jsonargparse.optionals import docstring_parser_support, set_docstring_parse_options
+from jsonargparse.util import unresolvable_import_paths
 
 
 responses_available = find_spec('responses') is not None
@@ -44,6 +46,12 @@ def mock_module(*args):
     import jsonargparse_tests
     with unittest.mock.patch.multiple(jsonargparse_tests, create=True, **{c.__name__: c for c in args}):
         yield __module__
+
+
+def doctest_mock_class_in_main(cls):
+    cls.__module__ = None
+    setattr(sys.modules['__main__'], cls.__name__, cls)
+    unresolvable_import_paths[cls] = f'__main__.{cls.__name__}'
 
 
 @contextmanager