+++ CHANGELOG.rst
@@ -12,7 +12,7 @@ The semantic versioning only considers the public API as described in
 paths are considered internals and can change in minor and patch releases.
 
 
-v4.18.0 (2022-11-??)
+v4.18.0 (2022-11-29)
 --------------------
 
 Added